refactor: service.js 파일에 도메인 로직 분리
