+++ src/components/Receipt/PurchaseNumberItem/index.js
@@ -3,7 +3,7 @@ import { v4 as uuidv4 } from 'uuid';
 
 import LotteryBall from '../../shared/LotteryBall';
 
-import chooseBallColor from '../../../utils/colorBall';
+import { chooseBallColor } from '../../service';
 
 import './style.scss';
